Refactor MessageMappingDataGrid
